+++ src/modules/scaler.rs
@@ -91,7 +91,7 @@ impl Module for Scaler {
         self.knobs[i] = val;
     }
 
-    fn step(&mut self, _time: f32, st: StepType, ins: &[f32]) -> Vec<f32> {
+    fn step(&mut self, _time: f64, st: StepType, ins: &[f32]) -> Vec<f32> {
         if st == StepType::Video {
             return vec![0.0];
         }
